agent-example/sick-leave-generator/agent/requirements.txt
crewai
crewai-tools
langchain-ollama
litellm
fpdf2